+++ src/test/java/com/kurkus/kusinsa/service/product/ProductStatusServiceTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class ProductStatusServiceTest {
+  
+}
